Revert "Update to 1.5"
This reverts commit b3d2448ac51607f4a5ee111805021baf2015aa99.

+++ Soldering-Station/frontend_190409.h
@@ -44,7 +44,7 @@ public:
 
 private:
     uint16_t display_timout =0;
-    uint16_t display_heat_blink=0;
+    uint8_t display_heat_blink=0;
 
 
 };
